Created audioProcess.ts module
